Refactor user/student/instructor service into separate modules
