+++ app/services/__init__.py
@@ -1,5 +1,5 @@
+from .user import *
 from .jwt_service import *
-from .user_service import *
 from .assignment_service import *
 from .submission_service import *
 from .course_service import *
