+++ app/services/user/user_service.py
@@ -1,0 +1,44 @@
+from typing import List
+from sqlalchemy.orm import Session
+from app.models import UserModel
+from app.schemas import RefreshTokenSchema
+from app.core.config import settings
+from app.core.utils.token_helper import TokenHelper
+from app.core.utils.auth_helper import PasswordHelper
+from app.core.middleware.authentication import CurrentUser
+from app.core.exceptions import (
+    PasswordDoesNotMatchException,
+    UserNotFoundException,
+    PasswordDoesNotMatchException
+)
+
+class UserService:
+    def __init__(self, session: Session):
+        self.session = session
+
+    async def get_user_by_id(self, id: int) -> UserModel:
+        user = self.session.query(UserModel).filter_by(id=id).first()
+        if user is None:
+            raise UserNotFoundException()
+        return user
+
+    async def get_user_by_onyen(self, onyen: str) -> UserModel:
+        user = self.session.query(UserModel).filter_by(onyen=onyen).first()
+        if user is None:
+            raise UserNotFoundException()
+        return user
+
+    async def login(self, onyen: str, password: str) -> RefreshTokenSchema:
+        user = self.session.query(UserModel).filter_by(onyen=onyen).first()
+        if not user:
+            raise UserNotFoundException()
+        if not PasswordHelper.verify_password(password, user.password):
+            raise PasswordDoesNotMatchException()
+
+        current_user = CurrentUser(id=user.id, onyen=user.onyen)
+
+        response = RefreshTokenSchema(
+            access_token=TokenHelper.encode(payload=current_user.dict(), expire_period=settings.ACCESS_TOKEN_EXPIRES_MINUTES),
+            refresh_token=TokenHelper.encode(payload={"sub": "refresh", **current_user.dict()}, expire_period=settings.REFRESH_TOKEN_EXPIRES_MINUTES)
+        )
+        return response
